BasicStroke caching

+++ Map/Line.java
@@ -8,9 +8,9 @@ public class Line {
 
 	public Vector start, stop;
 	public Color color;
-	public float width;
+	public int width;
 
-	public Line set(Vector start, Vector stop, Color color, float width) {
+	public Line set(Vector start, Vector stop, Color color, int width) {
 		this.start = start;
 		this.stop = stop;
 		this.color = color;
